Scope notification task JSON to its message handler
Committed-By-Agent: codex
