+++ StripeConnect/StripeConnect/Source/Internal/Webview/MessageHandlers/NotificationBanner/OpenNotificationBannerTaskMessageHandler.swift
@@ -6,6 +6,62 @@
 import Foundation
 
 final class OpenNotificationBannerTaskMessageHandler: ScriptMessageHandler<OpenNotificationBannerTaskMessageHandler.Payload> {
+    enum JSONValue: Codable, Equatable {
+        case array([JSONValue])
+        case bool(Bool)
+        case double(Double)
+        case integer(Int64)
+        case null
+        case object([String: JSONValue])
+        case string(String)
+
+        init(from decoder: Decoder) throws {
+            let container = try decoder.singleValueContainer()
+
+            if container.decodeNil() {
+                self = .null
+            } else if let value = try? container.decode(Bool.self) {
+                self = .bool(value)
+            } else if let value = try? container.decode(Int64.self) {
+                self = .integer(value)
+            } else if let value = try? container.decode(Double.self) {
+                self = .double(value)
+            } else if let value = try? container.decode(String.self) {
+                self = .string(value)
+            } else if let value = try? container.decode([String: JSONValue].self) {
+                self = .object(value)
+            } else if let value = try? container.decode([JSONValue].self) {
+                self = .array(value)
+            } else {
+                throw DecodingError.dataCorruptedError(
+                    in: container,
+                    debugDescription: "Unsupported JSON value"
+                )
+            }
+        }
+
+        func encode(to encoder: Encoder) throws {
+            var container = encoder.singleValueContainer()
+
+            switch self {
+            case .array(let value):
+                try container.encode(value)
+            case .bool(let value):
+                try container.encode(value)
+            case .double(let value):
+                try container.encode(value)
+            case .integer(let value):
+                try container.encode(value)
+            case .null:
+                try container.encodeNil()
+            case .object(let value):
+                try container.encode(value)
+            case .string(let value):
+                try container.encode(value)
+            }
+        }
+    }
+
     typealias Payload = [String: JSONValue]
 
     init(analyticsClient: ComponentAnalyticsClient,
